Move xbtop components into python module
Names like ReportPower or ReportMemory are too generic to be installed
into python path without module.
