+++ src/runtime_src/core/tools/xbtop/CMakeLists.txt
@@ -5,11 +5,12 @@ string(SUBSTRING "${LINUX_VERSION}" 0 1 major_version)
 set (PYTHON_3 TRUE)
 
 if (${PYTHON_3})
+  install(DIRECTORY
+    _xbtop
+    DESTINATION ${XRT_INSTALL_PYTHON_DIR} COMPONENT ${XRT_BASE_COMPONENT}
+    PATTERN "__pycache__" EXCLUDE)
+
   install(FILES
-    ReportDynamicRegions.py
-    ReportMemory.py
-    ReportPower.py
-    XBUtil.py
     xbtop.py
     DESTINATION ${XRT_INSTALL_PYTHON_DIR} COMPONENT ${XRT_BASE_COMPONENT})
